refactor: extract TelemetryTypes.swift from OperationalTelemetryRecorder.swift
Extracts 3 support types (OperationalTelemetryEvent struct, TelemetryEventName
struct, TelemetryEvent enum) from OperationalTelemetryRecorder.swift into a
new TelemetryTypes.swift.

Byte-identical move; SHA verified against origin/main.

Reduces OperationalTelemetryRecorder.swift from 134 to 93 lines. Only the
primary OperationalTelemetryRecorder struct remains.

Tests: 667 (unchanged).
